+++ org.eclipse.winery.repository.ui/src/app/wineryInterfaces/toscaComponent.ts
@@ -32,4 +32,8 @@ export class ToscaComponent {
             }
         }
     }
+
+    getQName(): string {
+        return '{' + this.namespace + '}' + this.localName;
+    }
 }